fix: 알림 상태 업데이트와 관련된 로직을 분리
